Refactored and cleaned code up. DRY baby! Fixed issues with enemy ship not being detected when building is struck. Fixed issue with buildings hit box. Updated invader transitions down the screen for smoother game play. Updated status bar to reflect score and remaining player lives with ship graphics. Updated players controls to capture left and right gestures. Added physics mechanics to all game's sprite movements for more more realistic transitions. Added a ready screen to display and allow player to prepare for battle. Reduced initial enemy bullet cap to 5. Increased initial menace interval to 1999 and added a constraint to keep it from becoming zero. Added lifeLost boolean to assist with correctly decrementing remaining life count. Added boolean hasLanded to help with displaying marauders landed. Added endGame method to assist in clearing screen and ending game play. Added a new save score activity to store high score initials and then transition player back to main activity. Utilized shared preferences to load and store high scores. Update high score page with new graphics and logic to display high scores. Updated star class with physics for smoother background movement.

+++ app/src/main/java/com/oomph/spacemarauders/Boom.java
@@ -7,17 +7,14 @@ import android.content.Context;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 
-/**
- * Created by Belal on 6/15/2016.
- */
 public class Boom {
 
     //bitmap object
     private Bitmap bitmap;
 
     //coordinate variables
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
     //constructor
     public Boom(Context context) {
@@ -34,11 +31,11 @@ public class Boom {
     }
 
     //setters for x and y to make it visible at the place of collision
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 
@@ -51,11 +48,11 @@ public class Boom {
         this.bitmap = bitmap;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
